Create speaeate component for `IndexSetCell`.
